requirements.txt
feedparser
pandas
pyyaml
python-dotenv
requests
